requirements.txt
http://www.djangoproject.com/download/1.7.b4/tarball/
django-suit
pika
mysql-python
xhtml2pdf
qrcode
